findx-api/requirements.txt
asgiref==3.5.0
Django==4.0.4
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
importlib-metadata==4.11.3
Markdown==3.3.6
PyJWT==2.3.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
zipp==3.8.0
